+++ laktory/dlt/__init__.py
@@ -5,6 +5,7 @@ try:
     from dlt import *
     from dlt import read as _read
     from dlt import read_stream as _read_stream
+    from dlt import apply_changes as _apply_changes
 except Exception:
     pass
 
@@ -93,6 +94,18 @@ def read_stream(*args, fmt="delta", **kwargs):
         return _read_stream(*args, **kwargs)
 
 
+def apply_changes(*args, table=None, **kwargs):
+    if is_debug():
+        if table is None:
+            return
+        df = table.read_source(spark=spark)
+        # TODO: Apply changes
+        logger.warning("Laktory does not currently support applying CDC changes. Returned DataFrame is CDC source.")
+        return df
+    else:
+        return _apply_changes(*args, **kwargs)
+
+
 # --------------------------------------------------------------------------- #
 # Mocks                                                                       #
 # --------------------------------------------------------------------------- #
@@ -172,3 +185,7 @@ if is_mocked():
             return wrapper
 
         return decorator
+
+
+    def create_streaming_table(*args, **kwargs):
+        pass