+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Release History
 
-## [0.0.15]
+## [0.0.16]
 **Breaking Change**
 - Refactored InitScript model into the more general WorkspaceFile
+
+## [0.0.15]
+**Added**
+- Support for CDC table data source
+- Support for SCD table
+
+**Updated**
+- Automatic catalog and schema assignation to source table from table and pipeline
 
 ## [0.0.14] - 2023-11-06
 **Fixed**